refactor: tranfer methods menu and randomIndexGenerator to class game
